feat: add easy loop syntax sugars
